requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
asgiref==3.6.0
    # via django
cachetools==5.2.0
    # via google-auth
certifi==2022.12.7
    # via requests
charset-normalizer==2.1.1
    # via requests
coreapi==2.3.3
    # via -r requirements.in
coreschema==0.0.4
    # via coreapi
django==4.1.4
    # via
    #   -r requirements.in
    #   django-debug-toolbar
    #   djangorestframework
django-crispy-forms==1.14.0
    # via -r requirements.in
django-debug-toolbar==3.8.1
    # via -r requirements.in
djangorestframework==3.14.0
    # via -r requirements.in
google-api-core[grpc]==2.11.0
    # via
    #   google-cloud-bigquery
    #   google-cloud-core
google-auth==2.15.0
    # via
    #   google-api-core
    #   google-cloud-core
google-cloud-bigquery==3.4.1
    # via -r requirements.in
google-cloud-core==2.3.2
    # via google-cloud-bigquery
google-crc32c==1.5.0
    # via google-resumable-media
google-resumable-media==2.4.0
    # via google-cloud-bigquery
googleapis-common-protos==1.57.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.51.1
    # via
    #   google-api-core
    #   google-cloud-bigquery
    #   grpcio-status
grpcio-status==1.51.1
    # via google-api-core
idna==3.4
    # via requests
itypes==1.2.0
    # via coreapi
jinja2==3.1.2
    # via coreschema
joblib==1.2.0
    # via -r requirements.in
markupsafe==2.1.1
    # via jinja2
mysqlclient==2.1.1
    # via -r requirements.in
packaging==21.3
    # via google-cloud-bigquery
proto-plus==1.22.1
    # via google-cloud-bigquery
protobuf==4.21.12
    # via
    #   google-api-core
    #   google-cloud-bigquery
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pyparsing==3.0.9
    # via packaging
python-dateutil==2.8.2
    # via google-cloud-bigquery
pytz==2022.7
    # via djangorestframework
requests==2.28.1
    # via
    #   coreapi
    #   google-api-core
    #   google-cloud-bigquery
    #   shopcloud-secrethub
rsa==4.9
    # via google-auth
shopcloud-secrethub==2.10.0
    # via -r requirements.in
six==1.16.0
    # via
    #   google-auth
    #   python-dateutil
sqlparse==0.4.3
    # via
    #   django
    #   django-debug-toolbar
uritemplate==4.1.1
    # via coreapi
urllib3==1.26.13
    # via requests
